LWDEV-6992 use big decimal in tests

+++ src/test/kotlin/com/lykke/matching/engine/services/ClientMultiLimitOrderTest.kt
@@ -26,6 +26,7 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Primary
 import org.springframework.test.annotation.DirtiesContext
 import org.springframework.test.context.junit4.SpringRunner
+import java.math.BigDecimal
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
@@ -95,11 +96,11 @@ class ClientMultiLimitOrderTest : AbstractTest() {
         report.orders.forEach {
             assertEquals(OrderStatus.InOrderBook.name, it.order.status)
         }
-        assertEquals(0.1, report.orders.first { it.order.externalId == "1" }.order.reservedLimitVolume)
-        assertEquals(0.2, report.orders.first { it.order.externalId == "2" }.order.reservedLimitVolume)
-        assertEquals(0.30000001, report.orders.first { it.order.externalId == "3" }.order.reservedLimitVolume)
-        assertEquals(950.0, report.orders.first { it.order.externalId == "4" }.order.reservedLimitVolume)
-        assertEquals(1800.0, report.orders.first { it.order.externalId == "5" }.order.reservedLimitVolume)
+        assertEquals(BigDecimal.valueOf(0.1), report.orders.first { it.order.externalId == "1" }.order.reservedLimitVolume)
+        assertEquals(BigDecimal.valueOf(0.2), report.orders.first { it.order.externalId == "2" }.order.reservedLimitVolume)
+        assertEquals(BigDecimal.valueOf(0.30000001), report.orders.first { it.order.externalId == "3" }.order.reservedLimitVolume)
+        assertEquals(BigDecimal.valueOf(950.0), report.orders.first { it.order.externalId == "4" }.order.reservedLimitVolume)
+        assertEquals(BigDecimal.valueOf(1800.0), report.orders.first { it.order.externalId == "5" }.order.reservedLimitVolume)
     }
 
     @Test
@@ -163,9 +164,9 @@ class ClientMultiLimitOrderTest : AbstractTest() {
         val buyOrderBook = orderBookQueue.first { it.isBuy }
         val sellOrderBook = orderBookQueue.first { !it.isBuy }
         assertEquals(2, buyOrderBook.prices.size)
-        assertEquals(9500.0, buyOrderBook.prices.first().price)
+        assertEquals(BigDecimal.valueOf(9500.0), buyOrderBook.prices.first().price)
         assertEquals(4, sellOrderBook.prices.size)
-        assertEquals(10000.0, sellOrderBook.prices.first().price)
+        assertEquals(BigDecimal.valueOf(10000.0), sellOrderBook.prices.first().price)
 
         assertEquals(1, clientsLimitOrdersQueue.size)
         assertEquals(0, trustedClientsLimitOrdersQueue.size)
@@ -212,7 +213,7 @@ class ClientMultiLimitOrderTest : AbstractTest() {
 
         val sellOrderBook = orderBookQueue.first { !it.isBuy }
         assertEquals(3, sellOrderBook.prices.size)
-        assertEquals(10000.0, sellOrderBook.prices.first().price)
+        assertEquals(BigDecimal.valueOf(10000.0), sellOrderBook.prices.first().price)
 
         assertBalance("Client1", "BTC", 1.0, 0.60000001)
 
@@ -385,8 +386,8 @@ class ClientMultiLimitOrderTest : AbstractTest() {
         assertOrderBookSize("BTCUSD", false, 1)
         assertOrderBookSize("BTCUSD", true, 1)
 
-        assertEquals(9000.0, genericLimitOrderService.getOrderBook("BTCUSD").getBidPrice())
-        assertEquals(9500.0, genericLimitOrderService.getOrderBook("BTCUSD").getAskPrice())
+        assertEquals(BigDecimal.valueOf(9000.0), genericLimitOrderService.getOrderBook("BTCUSD").getBidPrice())
+        assertEquals(BigDecimal.valueOf(9500.0), genericLimitOrderService.getOrderBook("BTCUSD").getAskPrice())
 
         assertBalance("Client1", "USD", 1355.0, 900.0)
         assertBalance("Client1", "BTC", 0.11, 0.0)
